chore: Centralizing types
