+++ components/custom/AddressInterface/AddNewAddress/ChosenResiduesInterface/types/index.ts
@@ -1,0 +1,7 @@
+import { Residue } from "@/components/custom/WasteManagementInterface/types";
+
+type ChosenResiduesInterfaceProps = {
+  residues: Residue[];
+};
+
+export { ChosenResiduesInterfaceProps };
