+++ components/custom/AddressInterface/RegisteredAddresses/RegisteredAddressCard/types/index.ts
@@ -1,0 +1,8 @@
+import { Address } from "../../../types";
+
+interface RegisteredAddressCardProps {
+  item: Address;
+  onSelect: (id: string) => void;
+  selected: boolean;
+}
+export { RegisteredAddressCardProps };
