+++ components/custom/UserInterface/components/WasteProducerStatsInterface/components/AccountSection/AccountSection.tsx
@@ -10,13 +10,7 @@ import {
 } from "lucide-react-native";
 import { BarChart } from "react-native-gifted-charts";
 import { EventDetail } from "@/context/WasteProducerStatsContext";
-import { User } from "@/types";
-
-interface AccountSectionProps {
-  user: User | null;
-  loading: boolean;
-  events: EventDetail[];
-}
+import { AccountSectionProps } from "../types";
 
 export const AccountSection: FC<AccountSectionProps> = ({
   user,
